set_inputs mask type int32

+++ examples/opensora_hpcai/scripts/train.py
@@ -672,7 +672,7 @@ def main(args):
     if (args.mode == 0) and (args.bucket_config is not None):
         video = ms.Tensor(shape=[None, None, 3, None, None], dtype=ms.float32)
         caption = ms.Tensor(shape=[None, args.model_max_length, 4096], dtype=ms.float32)
-        mask = ms.Tensor(shape=[None, args.model_max_length], dtype=ms.uint8)
+        mask = ms.Tensor(shape=[None, args.model_max_length], dtype=ms.int32)
         frames_mask = ms.Tensor(shape=[None, None], dtype=ms.bool_)
         # fmt: off
         num_frames = ms.Tensor(shape=[None, ], dtype=ms.float32)
